Move data checking from data inserter to data checker service
