+++ workers/loc.api/sync/data.inserter/data.checker/helpers/invert-sort.js
@@ -1,0 +1,11 @@
+'use strict'
+
+module.exports = (sortArr) => {
+  return sortArr.map(item => {
+    const _arr = [...item]
+
+    _arr[1] = item[1] > 0 ? -1 : 1
+
+    return _arr
+  })
+}
