+++ workers/loc.api/sync/data.inserter/helpers/index.js
@@ -7,23 +7,27 @@ const filterMethodCollMapByList = require(
   './filter-method-coll-map-by-list'
 )
 const {
-  invertSort,
-  filterMethodCollMap,
-  checkCollType,
-  compareElemsDbAndApi,
   normalizeApiData,
   getAuthFromDb,
   getAllowedCollsNames
 } = require('./utils')
+const {
+  checkCollType,
+  isInsertableArrObjTypeOfColl,
+  isUpdatableArrObjTypeOfColl,
+  isUpdatableArrTypeOfColl
+} = require('./check-coll-type')
+const getMethodArgMap = require('./get-method-arg-map')
 
 module.exports = {
   searchClosePriceAndSumAmount,
   filterMethodCollMapByList,
-  invertSort,
-  filterMethodCollMap,
-  checkCollType,
-  compareElemsDbAndApi,
   normalizeApiData,
   getAuthFromDb,
-  getAllowedCollsNames
+  getAllowedCollsNames,
+  checkCollType,
+  isInsertableArrObjTypeOfColl,
+  isUpdatableArrObjTypeOfColl,
+  isUpdatableArrTypeOfColl,
+  getMethodArgMap
 }
