+++ workers/loc.api/sync/data.inserter/helpers/utils.js
@@ -4,55 +4,6 @@ const {
   pick
 } = require('lodash')
 
-const invertSort = (sortArr) => {
-  return sortArr.map(item => {
-    const _arr = [...item]
-
-    _arr[1] = item[1] > 0 ? -1 : 1
-
-    return _arr
-  })
-}
-
-const filterMethodCollMap = (
-  methodCollMap,
-  isPublic
-) => {
-  return new Map([...methodCollMap].filter(([key, schema]) => {
-    const _isHidden = /^hidden:/i.test(schema.type)
-
-    if (_isHidden) return false
-
-    const _isPub = /^public:/i.test(schema.type)
-
-    return schema.hasNewData && (isPublic ? _isPub : !_isPub)
-  }))
-}
-
-const checkCollType = (
-  type,
-  coll,
-  isPublic
-) => {
-  const _pub = isPublic ? 'public:' : ''
-  const regExp = new RegExp(`^${_pub}${type}$`, 'i')
-
-  return regExp.test(coll.type)
-}
-
-const compareElemsDbAndApi = (
-  dateFieldName,
-  elDb,
-  elApi
-) => {
-  const _elDb = Array.isArray(elDb) ? elDb[0] : elDb
-  const _elApi = Array.isArray(elApi) ? elApi[0] : elApi
-
-  return (_elDb[dateFieldName] < _elApi[dateFieldName])
-    ? _elDb[dateFieldName]
-    : false
-}
-
 const normalizeApiData = (
   data = [],
   model,
@@ -133,10 +84,6 @@ const getAllowedCollsNames = (allowedColls) => {
 }
 
 module.exports = {
-  invertSort,
-  filterMethodCollMap,
-  checkCollType,
-  compareElemsDbAndApi,
   normalizeApiData,
   getAuthFromDb,
   getAllowedCollsNames
