+++ workers/loc.api/sync/data.inserter/index.js
@@ -1,7 +1,6 @@
 'use strict'
 
 const EventEmitter = require('events')
-const moment = require('moment')
 const {
   isEmpty,
   cloneDeep
@@ -21,13 +20,13 @@ const {
 const TYPES = require('../../di/types')
 const {
   filterMethodCollMapByList,
-  invertSort,
-  filterMethodCollMap,
-  checkCollType,
-  compareElemsDbAndApi,
+  isInsertableArrObjTypeOfColl,
+  isUpdatableArrObjTypeOfColl,
+  isUpdatableArrTypeOfColl,
   normalizeApiData,
   getAuthFromDb,
-  getAllowedCollsNames
+  getAllowedCollsNames,
+  getMethodArgMap
 } = require('./helpers')
 const DataInserterHook = require('./hooks/data.inserter.hook')
 const {
@@ -48,7 +47,6 @@ class DataInserter extends EventEmitter {
     syncSchema,
     TABLES_NAMES,
     ALLOWED_COLLS,
-    currencyConverter,
     FOREX_SYMBS,
     authenticator,
     convertCurrencyHook,
@@ -63,7 +61,6 @@ class DataInserter extends EventEmitter {
     this.syncSchema = syncSchema
     this.TABLES_NAMES = TABLES_NAMES
     this.ALLOWED_COLLS = ALLOWED_COLLS
-    this.currencyConverter = currencyConverter
     this.FOREX_SYMBS = FOREX_SYMBS
     this.authenticator = authenticator
     this.convertCurrencyHook = convertCurrencyHook
@@ -79,8 +76,8 @@ class DataInserter extends EventEmitter {
 
     this.convertTo = 'USD'
 
-    this._candlesTimeframe = '1D'
-    this._candlesSection = 'hist'
+    this.candlesTimeframe = '1D'
+    this.candlesSection = 'hist'
   }
 
   init (syncColls = this.ALLOWED_COLLS.ALL) {
@@ -104,6 +101,11 @@ class DataInserter extends EventEmitter {
       this.convertCurrencyHook,
       this.recalcSubAccountLedgersBalancesHook
     ])
+    this.dataChecker.init({
+      convertTo: this.convertTo,
+      candlesTimeframe: this.candlesTimeframe,
+      candlesSection: this.candlesSection
+    })
   }
 
   addAsyncProgressHandler (handler) {
@@ -197,7 +199,9 @@ class DataInserter extends EventEmitter {
   }
 
   async insertNewPublicDataToDb (prevProgress) {
-    const methodCollMap = await this.checkNewPublicData()
+    // TODO:
+    const methodCollMap = await this.dataChecker
+      .checkNewPublicData()
     const size = methodCollMap.size
 
     let count = 0
@@ -227,14 +231,24 @@ class DataInserter extends EventEmitter {
       return
     }
 
-    const methodCollMap = await this.checkNewData(auth)
+    // TODO:
+    const methodCollMap = await this.dataChecker
+      .checkNewData(auth)
     const size = this._methodCollMap.size
 
     let count = 0
     let progress = 0
 
     for (const [method, item] of methodCollMap) {
-      const args = this._getMethodArgMap(method, auth, 10000000, item.start)
+      const { start } = item
+      const args = this._getMethodArgMap(
+        method,
+        {
+          auth,
+          limit: 10000000,
+          start
+        }
+      )
 
       await this._insertApiDataArrObjTypeToDb(
         args,
@@ -255,375 +269,6 @@ class DataInserter extends EventEmitter {
     return progress
   }
 
-  async checkNewData (auth) {
-    const methodCollMap = this._getMethodCollMap()
-
-    await this._checkNewDataArrObjType(auth, methodCollMap)
-
-    return filterMethodCollMap(methodCollMap)
-  }
-
-  async checkNewPublicData () {
-    const methodCollMap = this._getMethodCollMap()
-
-    await this._checkNewDataPublicArrObjType(methodCollMap)
-
-    return filterMethodCollMap(methodCollMap, true)
-  }
-
-  async _checkNewDataPublicArrObjType (methodCollMap) {
-    for (const [method, schema] of methodCollMap) {
-      if (!this._isInsertableArrObjTypeOfColl(schema, true)) {
-        continue
-      }
-      if (
-        schema.name === this.ALLOWED_COLLS.PUBLIC_TRADES ||
-        schema.name === this.ALLOWED_COLLS.TICKERS_HISTORY
-      ) {
-        schema.hasNewData = false
-
-        await this._checkNewConfigurablePublicData(method, schema)
-
-        continue
-      }
-      if (schema.name === this.ALLOWED_COLLS.CANDLES) {
-        schema.hasNewData = false
-
-        await this._checkNewCandlesData(method, schema)
-        await this._checkNewConfigurablePublicData(method, schema)
-
-        continue
-      }
-
-      await this._checkItemNewDataArrObjType(method, schema)
-    }
-  }
-
-  async _checkNewConfigurablePublicData (method, schema) {
-    const {
-      confName,
-      symbolFieldName,
-      timeframeFieldName,
-      dateFieldName,
-      name,
-      sort
-    } = { ...schema }
-    const groupResBy = (
-      timeframeFieldName &&
-      typeof timeframeFieldName === 'string'
-    )
-      ? ['symbol', 'timeframe']
-      : ['symbol']
-    const publicСollsСonf = await this.dao.getElemsInCollBy(
-      this.TABLES_NAMES.PUBLIC_COLLS_CONF,
-      {
-        filter: { confName },
-        subQuery: { sort: [['start', 1]] },
-        groupResBy
-      }
-    )
-
-    if (isEmpty(publicСollsСonf)) {
-      return
-    }
-
-    const params = name === this.ALLOWED_COLLS.CANDLES
-      ? {
-        section: this._candlesSection,
-        notThrowError: true,
-        notCheckNextPage: true
-      }
-      : {
-        notThrowError: true,
-        notCheckNextPage: true
-      }
-
-    for (const confs of publicСollsСonf) {
-      const {
-        symbol,
-        start,
-        timeframe
-      } = confs
-      const timeframeParam = (
-        timeframe &&
-        typeof timeframe === 'string'
-      )
-        ? { timeframe }
-        : {}
-      const args = this._getMethodArgMap(
-        method,
-        {},
-        1,
-        0,
-        Date.now(),
-        {
-          ...params,
-          ...timeframeParam,
-          symbol
-        }
-      )
-      const timeframeFilter = (
-        timeframe &&
-        typeof timeframe === 'string' &&
-        timeframeFieldName &&
-        typeof timeframeFieldName === 'string'
-      )
-        ? { [timeframeFieldName]: timeframe }
-        : {}
-      const filter = {
-        ...timeframeFilter,
-        [symbolFieldName]: symbol
-      }
-      const lastElemFromDb = await this.dao.getElemInCollBy(
-        name,
-        filter,
-        sort
-      )
-      const { res: lastElemFromApi } = await this._getDataFromApi(
-        method,
-        args,
-        true
-      )
-
-      if (
-        isEmpty(lastElemFromApi) ||
-        (
-          Array.isArray(lastElemFromApi) &&
-          lastElemFromApi[0][symbolFieldName] &&
-          typeof lastElemFromApi[0][symbolFieldName] === 'string' &&
-          lastElemFromApi[0][symbolFieldName] !== symbol
-        )
-      ) {
-        continue
-      }
-      if (isEmpty(lastElemFromDb)) {
-        schema.hasNewData = true
-        this._pushConfigurablePublicDataStartConf(
-          schema,
-          symbol,
-          { currStart: start },
-          timeframe
-        )
-
-        continue
-      }
-
-      const lastDateInDb = compareElemsDbAndApi(
-        dateFieldName,
-        lastElemFromDb,
-        lastElemFromApi
-      )
-
-      const startConf = {
-        baseStartFrom: null,
-        baseStartTo: null,
-        currStart: null
-      }
-
-      if (lastDateInDb) {
-        schema.hasNewData = true
-        startConf.currStart = lastDateInDb + 1
-      }
-
-      const firstElemFromDb = await this.dao.getElemInCollBy(
-        name,
-        filter,
-        invertSort(sort)
-      )
-
-      if (!isEmpty(firstElemFromDb)) {
-        const isChangedBaseStart = compareElemsDbAndApi(
-          dateFieldName,
-          { [dateFieldName]: start },
-          firstElemFromDb
-        )
-
-        if (isChangedBaseStart) {
-          schema.hasNewData = true
-          startConf.baseStartFrom = start
-          startConf.baseStartTo = firstElemFromDb[dateFieldName] - 1
-        }
-      }
-
-      this._pushConfigurablePublicDataStartConf(
-        schema,
-        symbol,
-        startConf,
-        timeframe
-      )
-    }
-  }
-
-  _pushConfigurablePublicDataStartConf (
-    schema,
-    symbol,
-    startConf = {},
-    timeframe
-  ) {
-    const {
-      baseStartFrom,
-      baseStartTo,
-      currStart
-    } = { ...startConf }
-
-    const currStartConfArr = schema.start
-      .find(([symb, conf, tFrame]) => (
-        symb === symbol &&
-        (
-          !timeframe ||
-          tFrame === timeframe
-        )
-      ))
-
-    if (!Array.isArray(currStartConfArr)) {
-      schema.start.push([
-        symbol,
-        {
-          baseStartFrom,
-          baseStartTo,
-          currStart
-        },
-        timeframe
-      ])
-
-      return
-    }
-
-    const currStartConf = { ...currStartConfArr[1] }
-    const _startConf = {
-      baseStartFrom: (
-        Number.isInteger(currStartConf.baseStartFrom) &&
-        (
-          !Number.isInteger(baseStartFrom) ||
-          currStartConf.baseStartFrom < baseStartFrom
-        )
-      )
-        ? currStartConf.baseStartFrom
-        : baseStartFrom,
-      baseStartTo: (
-        Number.isInteger(currStartConf.baseStartTo) &&
-        (
-          !Number.isInteger(baseStartTo) ||
-          currStartConf.baseStartTo > baseStartTo
-        )
-      )
-        ? currStartConf.baseStartTo
-        : baseStartTo,
-      currStart: (
-        Number.isInteger(currStartConf.currStart) &&
-        (
-          !Number.isInteger(currStart) ||
-          currStartConf.currStart < currStart
-        )
-      )
-        ? currStartConf.currStart
-        : currStart
-    }
-
-    currStartConfArr[1] = {
-      ...currStartConf,
-      ..._startConf
-    }
-  }
-
-  async _checkItemNewDataArrObjType (
-    method,
-    schema,
-    auth
-  ) {
-    schema.hasNewData = false
-
-    const args = this._getMethodArgMap(method, auth, 1)
-    args.params.notThrowError = true
-    args.params.notCheckNextPage = true
-
-    const { _id, subUser } = { ...auth }
-    const { _id: subUserId } = { ...subUser }
-    const hasSubUserIdField = (
-      schema.model &&
-      typeof schema.model === 'object' &&
-      typeof schema.model.subUserId === 'string' &&
-      Number.isInteger(subUserId)
-    )
-    const subUserIdFilter = hasSubUserIdField
-      ? { $eq: { subUserId } }
-      : {}
-    const lastElemFromDb = await this.dao.getElemInCollBy(
-      schema.name,
-      {
-        user_id: _id,
-        ...subUserIdFilter
-      },
-      schema.sort
-    )
-    const { res: lastElemFromApi } = await this._getDataFromApi(
-      method,
-      args,
-      true
-    )
-
-    if (isEmpty(lastElemFromApi)) {
-      return
-    }
-
-    if (isEmpty(lastElemFromDb)) {
-      schema.hasNewData = true
-      schema.start = 0
-      return
-    }
-
-    const lastDateInDb = compareElemsDbAndApi(
-      schema.dateFieldName,
-      lastElemFromDb,
-      lastElemFromApi
-    )
-
-    if (lastDateInDb) {
-      schema.hasNewData = true
-      schema.start = lastDateInDb + 1
-    }
-  }
-
-  async _checkNewDataArrObjType (auth, methodCollMap) {
-    for (const [method, item] of methodCollMap) {
-      if (!this._isInsertableArrObjTypeOfColl(item)) {
-        continue
-      }
-
-      await this._checkItemNewDataArrObjType(
-        method,
-        item,
-        auth
-      )
-    }
-
-    return methodCollMap
-  }
-
-  _isInsertableArrObjTypeOfColl (coll, isPublic) {
-    return checkCollType(
-      'insertable:array:objects',
-      coll,
-      isPublic
-    )
-  }
-
-  _isUpdatableArrObjTypeOfColl (coll, isPublic) {
-    return checkCollType(
-      'updatable:array:objects',
-      coll,
-      isPublic
-    )
-  }
-
-  _isUpdatableArrTypeOfColl (coll, isPublic) {
-    return checkCollType(
-      'updatable:array',
-      coll,
-      isPublic
-    )
-  }
-
   _getDataFromApi (methodApi, args, isCheckCall) {
     if (!this.apiMiddleware.hasMethod(methodApi)) {
       throw new FindMethodError()
@@ -641,7 +286,7 @@ class DataInserter extends EventEmitter {
     methodApi,
     schema
   ) {
-    if (!this._isInsertableArrObjTypeOfColl(schema, true)) {
+    if (!isInsertableArrObjTypeOfColl(schema, true)) {
       return
     }
 
@@ -656,7 +301,7 @@ class DataInserter extends EventEmitter {
         const addApiParams = name === this.ALLOWED_COLLS.CANDLES
           ? {
             timeframe,
-            section: this._candlesSection
+            section: this.candlesSection
           }
           : {}
 
@@ -690,10 +335,11 @@ class DataInserter extends EventEmitter {
     ) {
       const args = this._getMethodArgMap(
         methodApi,
-        null,
-        10000000,
-        dates.baseStartFrom,
-        dates.baseStartTo
+        {
+          limit: 10000000,
+          start: dates.baseStartFrom,
+          end: dates.baseStartTo
+        }
       )
       args.params = {
         ...args.params,
@@ -706,9 +352,10 @@ class DataInserter extends EventEmitter {
     if (Number.isInteger(dates.currStart)) {
       const args = this._getMethodArgMap(
         methodApi,
-        null,
-        10000000,
-        dates.currStart
+        {
+          limit: 10000000,
+          start: dates.currStart
+        }
       )
       args.params = {
         ...args.params,
@@ -726,7 +373,7 @@ class DataInserter extends EventEmitter {
     schema,
     isPublic
   ) {
-    if (!this._isInsertableArrObjTypeOfColl(schema, isPublic)) {
+    if (!isInsertableArrObjTypeOfColl(schema, isPublic)) {
       return
     }
 
@@ -834,7 +481,7 @@ class DataInserter extends EventEmitter {
     methodApi,
     schema
   ) {
-    if (!this._isUpdatableArrTypeOfColl(schema, true)) {
+    if (!isUpdatableArrTypeOfColl(schema, true)) {
       return
     }
 
@@ -843,7 +490,9 @@ class DataInserter extends EventEmitter {
       field
     } = schema
 
-    const args = this._getMethodArgMap(methodApi, {}, null, null, null)
+    const args = this._getMethodArgMap(
+      methodApi,
+      { start: null, end: null })
     const elemsFromApi = await this._getDataFromApi(methodApi, args)
 
     if (
@@ -893,16 +542,16 @@ class DataInserter extends EventEmitter {
       for (const type of syncingTypes) {
         const args = this._getMethodArgMap(
           methodApi,
-          {},
-          null,
-          null,
-          null,
           {
-            symbol,
-            filter: {
-              $gte: { [dateFieldName]: start }
-            },
-            type
+            start: null,
+            end: null,
+            params: {
+              symbol,
+              filter: {
+                $gte: { [dateFieldName]: start }
+              },
+              type
+            }
           }
         )
         const apiRes = await this._getDataFromApi(methodApi, args)
@@ -946,7 +595,7 @@ class DataInserter extends EventEmitter {
     methodApi,
     schema
   ) {
-    if (!this._isUpdatableArrObjTypeOfColl(schema, true)) {
+    if (!isUpdatableArrObjTypeOfColl(schema, true)) {
       return
     }
 
@@ -967,10 +616,7 @@ class DataInserter extends EventEmitter {
 
     const args = this._getMethodArgMap(
       methodApi,
-      {},
-      null,
-      null,
-      null
+      { start: null, end: null }
     )
     const apiRes = await this._getDataFromApi(methodApi, args)
     const elemsFromApi = (
@@ -1005,283 +651,15 @@ class DataInserter extends EventEmitter {
     }
   }
 
-  _getMethodArgMap (
-    method,
-    reqAuth,
-    reqLimit,
-    start = 0,
-    end = Date.now(),
-    params
-  ) {
-    const limit = reqLimit !== null
-      ? reqLimit
-      : this._methodCollMap.get(method).maxLimit
-
-    const { apiKey = '', apiSecret = '', subUser } = { ...reqAuth }
-    const {
-      apiKey: subUserApiKey,
-      apiSecret: subUserApiSecret
-    } = { ...subUser }
-    const auth = (
-      subUserApiKey &&
-      typeof subUserApiKey === 'string' &&
-      subUserApiSecret &&
-      typeof subUserApiSecret === 'string'
+  _getMethodArgMap (method, opts) {
+    return getMethodArgMap(
+      this._methodCollMap.get(method),
+      opts
     )
-      ? {
-        apiKey: subUserApiKey,
-        apiSecret: subUserApiSecret,
-        session: reqAuth
-      }
-      : { apiKey, apiSecret, session: reqAuth }
-
-    return {
-      auth,
-      params: {
-        ...params,
-        limit,
-        end,
-        start
-      }
-    }
   }
 
   _getMethodCollMap () {
     return new Map(this._methodCollMap)
-  }
-
-  async _checkNewCandlesData (
-    method,
-    schema
-  ) {
-    const {
-      symbolFieldName,
-      timeframeFieldName,
-      dateFieldName,
-      name,
-      sort
-    } = { ...schema }
-
-    const lastElemLedgers = await this.dao.getElemInCollBy(
-      this.ALLOWED_COLLS.LEDGERS,
-      { $not: { currency: this.FOREX_SYMBS } },
-      [['mts', 1]]
-    )
-
-    if (
-      !lastElemLedgers ||
-      typeof lastElemLedgers !== 'object' ||
-      !Number.isInteger(lastElemLedgers.mts)
-    ) {
-      return
-    }
-
-    const uniqueLedgersSymbs = await this.dao.getElemsInCollBy(
-      this.ALLOWED_COLLS.LEDGERS,
-      {
-        filter: { $not: { currency: this.FOREX_SYMBS } },
-        isDistinct: true,
-        projection: ['currency']
-      }
-    )
-
-    if (
-      !Array.isArray(uniqueLedgersSymbs) ||
-      uniqueLedgersSymbs.length === 0
-    ) {
-      return
-    }
-
-    const currenciesSynonymous = await this.currencyConverter
-      .getCurrenciesSynonymous()
-
-    const uniqueSymbs = uniqueLedgersSymbs.reduce((accum, ledger) => {
-      const { currency } = { ...ledger }
-
-      if (!currency) {
-        return accum
-      }
-
-      accum.push(currency)
-
-      const synonymous = currenciesSynonymous.get(currency)
-
-      if (!synonymous) {
-        return accum
-      }
-
-      const uniqueSynonymous = synonymous
-        .filter(([syn]) => (
-          accum.every((symb) => symb !== syn)
-        ))
-        .map(([syn]) => syn)
-
-      accum.push(...uniqueSynonymous)
-
-      return accum
-    }, [])
-
-    const _collСonfig = uniqueSymbs.map((currency) => {
-      const _currency = typeof currency === 'string'
-        ? currency.replace(/F0$/i, '')
-        : currency
-      const separator = (
-        typeof _currency === 'string' &&
-        _currency.length > 3
-      )
-        ? ':'
-        : ''
-
-      return {
-        symbol: `t${_currency}${separator}${this.convertTo}`,
-        start: lastElemLedgers.mts
-      }
-    })
-    const collСonfig = this.FOREX_SYMBS.reduce((accum, convertTo) => {
-      const _symb = `tBTC${convertTo}`
-
-      if (accum.every(({ symbol }) => symbol !== _symb)) {
-        accum.push({
-          symbol: _symb,
-          start: lastElemLedgers.mts
-        })
-      }
-
-      return accum
-    }, _collСonfig)
-
-    for (const { symbol, start: configStart } of collСonfig) {
-      const mtsMoment = moment.utc(configStart)
-        .add(-1, 'days')
-        .valueOf()
-      const _start = configStart
-        ? mtsMoment
-        : configStart
-      const params = {
-        timeframe: this._candlesTimeframe,
-        section: this._candlesSection,
-        notThrowError: true,
-        notCheckNextPage: true,
-        symbol
-      }
-      const _argsForLastElem = this._getMethodArgMap(method, {}, 1)
-      const argsForLastElem = {
-        ..._argsForLastElem,
-        params: {
-          ..._argsForLastElem.params,
-          ...params
-        }
-      }
-      const _argsForReceivingStart = this._getMethodArgMap(
-        method,
-        {},
-        1,
-        0,
-        _start
-      )
-      const argsForReceivingStart = {
-        ..._argsForReceivingStart,
-        params: {
-          ..._argsForReceivingStart.params,
-          ...params
-        }
-      }
-
-      const filter = {
-        [symbolFieldName]: symbol,
-        [timeframeFieldName]: this._candlesTimeframe
-      }
-      const lastElemFromDb = await this.dao.getElemInCollBy(
-        name,
-        filter,
-        sort
-      )
-      const {
-        res: lastElemFromApi
-      } = await this._getDataFromApi(method, argsForLastElem)
-      const {
-        res: startElemFromApi
-      } = await this._getDataFromApi(method, argsForReceivingStart)
-
-      if (
-        isEmpty(lastElemFromApi) ||
-        (
-          Array.isArray(lastElemFromApi) &&
-          lastElemFromApi[0][symbolFieldName] &&
-          typeof lastElemFromApi[0][symbolFieldName] === 'string' &&
-          lastElemFromApi[0][symbolFieldName] !== symbol
-        )
-      ) {
-        continue
-      }
-
-      const start = (
-        Array.isArray(startElemFromApi) &&
-        startElemFromApi[startElemFromApi.length - 1] &&
-        typeof startElemFromApi[startElemFromApi.length - 1] === 'object' &&
-        Number.isInteger(
-          startElemFromApi[startElemFromApi.length - 1][dateFieldName]
-        )
-      )
-        ? startElemFromApi[startElemFromApi.length - 1][dateFieldName]
-        : _start
-
-      if (isEmpty(lastElemFromDb)) {
-        schema.hasNewData = true
-        this._pushConfigurablePublicDataStartConf(
-          schema,
-          symbol,
-          { currStart: start },
-          this._candlesTimeframe
-        )
-
-        continue
-      }
-
-      const lastDateInDb = compareElemsDbAndApi(
-        dateFieldName,
-        lastElemFromDb,
-        lastElemFromApi
-      )
-
-      const startConf = {
-        baseStartFrom: null,
-        baseStartTo: null,
-        currStart: null
-      }
-
-      if (lastDateInDb) {
-        schema.hasNewData = true
-        startConf.currStart = lastDateInDb + 1
-      }
-
-      const firstElemFromDb = await this.dao.getElemInCollBy(
-        name,
-        filter,
-        invertSort(sort)
-      )
-
-      if (!isEmpty(firstElemFromDb)) {
-        const isChangedBaseStart = compareElemsDbAndApi(
-          dateFieldName,
-          { [dateFieldName]: start },
-          firstElemFromDb
-        )
-
-        if (isChangedBaseStart) {
-          schema.hasNewData = true
-          startConf.baseStartFrom = start
-          startConf.baseStartTo = firstElemFromDb[dateFieldName] - 1
-        }
-      }
-
-      this._pushConfigurablePublicDataStartConf(
-        schema,
-        symbol,
-        startConf,
-        this._candlesTimeframe
-      )
-    }
   }
 }
 
@@ -1292,11 +670,10 @@ decorate(inject(TYPES.ApiMiddleware), DataInserter, 2)
 decorate(inject(TYPES.SyncSchema), DataInserter, 3)
 decorate(inject(TYPES.TABLES_NAMES), DataInserter, 4)
 decorate(inject(TYPES.ALLOWED_COLLS), DataInserter, 5)
-decorate(inject(TYPES.CurrencyConverter), DataInserter, 6)
-decorate(inject(TYPES.FOREX_SYMBS), DataInserter, 7)
-decorate(inject(TYPES.Authenticator), DataInserter, 8)
-decorate(inject(TYPES.ConvertCurrencyHook), DataInserter, 9)
-decorate(inject(TYPES.RecalcSubAccountLedgersBalancesHook), DataInserter, 10)
-decorate(inject(TYPES.DataChecker), DataInserter, 11)
+decorate(inject(TYPES.FOREX_SYMBS), DataInserter, 6)
+decorate(inject(TYPES.Authenticator), DataInserter, 7)
+decorate(inject(TYPES.ConvertCurrencyHook), DataInserter, 8)
+decorate(inject(TYPES.RecalcSubAccountLedgersBalancesHook), DataInserter, 9)
+decorate(inject(TYPES.DataChecker), DataInserter, 10)
 
 module.exports = DataInserter
